+++ lib/cubits/invoice.rb
@@ -1,6 +1,6 @@
 module Cubits
   class Invoice < Resource
     path '/api/v1/invoices'
-    expose_methods :create, :find, :reload
+    expose_methods :create, :find, :reload, :from_callback
   end # class Invoice
 end